refactor lp07 so that it can be used by lp41
