+++ tests/unit/libpluto/lp07-orient/orienttest_head.c
@@ -1,0 +1,42 @@
+#define LEAK_DETECTIVE
+#define AGGRESSIVE 1
+#define XAUTH
+#define MODECFG
+#define DEBUG 1
+#define PRINT_SA_DEBUG 1
+#define USE_KEYRR 1
+
+#include "constants.h"
+#include "oswalloc.h"
+#include "whack.h"
+#include "../../programs/pluto/rcv_whack.h"
+
+#include "../../programs/pluto/connections.c"
+
+#include "whackmsgtestlib.c"
+#include "seam_timer.c"
+#include "seam_fakevendor.c"
+#include "seam_pending.c"
+#include "seam_ikev1.c"
+#include "seam_crypt.c"
+#include "seam_kernel.c"
+#include "seam_rnd.c"
+#include "seam_log.c"
+#include "seam_xauth.c"
+#include "seam_host_parker.c"
+#include "seam_terminate.c"
+#include "seam_x509.c"
+#include "seam_spdbstruct.c"
+#include "seam_demux.c"
+#include "seam_whack.c"
+#include "seam_initiate.c"
+#include "seam_keys.c"
+#include "seam_exitlog.c"
+#include "seam_dnskey.c"
+#include "seam_natt.c"
+
+#include "seam_gi_sha1.c"
+#include "seam_finish.c"
+
+u_int8_t reply_buffer[MAX_OUTPUT_UDP_SIZE];
+
