+++ tests/unit/libpluto/lp07-orient/output.txt
@@ -17,15 +17,20 @@ RC=0 "parker1--jj2": fd68:c9f9:4157:2:0:1::/96===192.168.1.1[@parker01.emmjay.cr
 RC=0 "parker1--jj2":     myip=unset; hisip=unset;
 RC=0 "parker1--jj2":   ike_life: 3600s; ipsec_life: 28800s; rekey_margin: 540s; rekey_fuzz: 100%; keyingtries: 0
 RC=0 "parker1--jj2":   policy: RSASIG+ENCRYPT+TUNNEL+PFS+IKEv2ALLOW+IKEv2Init+SAREFTRACK; prio: 64,96; interface: eth0; kind=CK_PERMANENT
+RC=0 "ikev2:parker--jj": fd68:c9f9:4157:2:0:1::/96===%any[@parker01.emmjay.credil.org]...132.213.238.7[@jamesjohnson.emmjay.credil.org]===fd68:c9f9:4157::/64; unrouted; eroute owner: #0
+RC=0 "ikev2:parker--jj":     myip=unset; hisip=unset;
+RC=0 "ikev2:parker--jj":   ike_life: 3600s; ipsec_life: 28800s; rekey_margin: 540s; rekey_fuzz: 100%; keyingtries: 0
+RC=0 "ikev2:parker--jj":   policy: RSASIG+ENCRYPT+TUNNEL+PFS+IKEv2ALLOW+IKEv2Init+SAREFTRACK; prio: 64,96; interface: eth0; kind=CK_PERMANENT
 ./orienttest deleting connection
+./orienttest leak: 2 * keep id name, item size: X
+./orienttest leak: ID host_pair, item size: X
+./orienttest leak: host_pair, item size: X
+./orienttest leak: host ip, item size: X
 ./orienttest leak: keep id name, item size: X
 ./orienttest leak: host ip, item size: X
 ./orienttest leak: keep id name, item size: X
 ./orienttest leak: connection name, item size: X
 ./orienttest leak: struct connection, item size: X
-./orienttest leak: 2 * keep id name, item size: X
-./orienttest leak: ID host_pair, item size: X
-./orienttest leak: host_pair, item size: X
 ./orienttest leak: policies path, item size: X
 ./orienttest leak: ocspcerts path, item size: X
 ./orienttest leak: aacerts path, item size: X
@@ -41,3 +46,4 @@ RC=0 "parker1--jj2":   policy: RSASIG+ENCRYPT+TUNNEL+PFS+IKEv2ALLOW+IKEv2Init+SA
 ./orienttest leak detective found Z leaks, total size X
 Pre-amble (offset: X): #!-pluto-whack-file- recorded on FOO
 processing parker1--jj2
+processing ikev2:parker--jj
