+++ tests/unit/libpluto/lp40-orientafterprivate/Makefile.testcase
@@ -1,6 +1,7 @@
 SAMPLEDIR=../samples
 WHACKFILE=${OUTPUTS}/ikev2client.record.${ARCH}
 ENDNAME=parker
-UNITTESTARGS=${WHACKFILE} parker1--jj2 ikev2:parker--jj
+CONNNAME=parker1--jj2 ikev2:parker--jj
+UNITTESTARGS=${WHACKFILE} ${CONNNAME}
 
 TESTNAME=orienttest
